+++ internal/mutator/executor_test.go
@@ -30,7 +30,7 @@ import (
 
 	"github.com/go-gremlins/gremlins/internal/mutant"
 	"github.com/go-gremlins/gremlins/internal/mutator"
-	"github.com/go-gremlins/gremlins/internal/mutator/internal/workerpool"
+	"github.com/go-gremlins/gremlins/internal/mutator/workerpool"
 
 	"github.com/go-gremlins/gremlins/internal/configuration"
 	"github.com/go-gremlins/gremlins/internal/gomodule"
