+++ internal/mutator/workerpool/workerpool_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	"github.com/go-gremlins/gremlins/internal/mutant"
-	"github.com/go-gremlins/gremlins/internal/mutator/internal/workerpool"
+	"github.com/go-gremlins/gremlins/internal/mutator/workerpool"
 
 	"github.com/go-gremlins/gremlins/internal/configuration"
 )
